+++ website/docusaurus.config.js
@@ -1,4 +1,4 @@
-/* eslint-disable global-require,import/no-extraneous-dependencies */
+/* eslint-disable global-require */
 const { externalLinkProcessor } = require('./tools/utils/externalLink');
 
 const packages = [
@@ -34,7 +34,7 @@ const packagesOrder = [
 
 /** @type {Partial<import('@docusaurus/types').DocusaurusConfig>} */
 module.exports = {
-    title: 'Crawlee · Build reliable crawlers. Fast.',
+    title: 'Crawlee for JavaScript · Build reliable crawlers. Fast.',
     tagline: 'Build reliable crawlers. Fast.',
     url: 'https://crawlee.dev',
     baseUrl: '/',
@@ -66,6 +66,7 @@ module.exports = {
                     showLastUpdateAuthor: true,
                     showLastUpdateTime: true,
                     path: '../docs',
+                    routeBasePath: 'js/docs',
                     sidebarPath: './sidebars.js',
                     rehypePlugins: [externalLinkProcessor],
                     disableVersioning: !!process.env.CRAWLEE_DOCS_FAST,
@@ -107,6 +108,7 @@ module.exports = {
                 projectRoot: `${__dirname}/..`,
                 changelogs: true,
                 readmes: true,
+                routeBasePath: 'js/api',
                 sortPackages: (a, b) => {
                     return packagesOrder.indexOf(a.packageName) - packagesOrder.indexOf(b.packageName);
                 },
@@ -121,33 +123,33 @@ module.exports = {
             {
                 redirects: [
                     {
-                        from: '/docs',
-                        to: '/docs/quick-start',
+                        from: '/js/docs',
+                        to: '/js/docs/quick-start',
                     },
                     {
-                        from: '/docs/next',
-                        to: '/docs/next/quick-start',
+                        from: '/js/docs/next',
+                        to: '/js/docs/next/quick-start',
                     },
                     {
-                        from: '/docs/guides/environment-variables',
-                        to: '/docs/guides/configuration',
+                        from: '/js/docs/guides/environment-variables',
+                        to: '/js/docs/guides/configuration',
                     },
                     {
-                        from: '/docs/guides/getting-started',
-                        to: '/docs/introduction',
+                        from: '/js/docs/guides/getting-started',
+                        to: '/js/docs/introduction',
                     },
                     {
-                        from: '/docs/guides/apify-platform',
-                        to: '/docs/deployment/apify-platform',
+                        from: '/js/docs/guides/apify-platform',
+                        to: '/js/docs/deployment/apify-platform',
                     },
                 ],
-                createRedirects(existingPath) {
-                    if (!existingPath.endsWith('/')) {
-                        return `${existingPath}/`;
-                    }
-
-                    return undefined; // Return a falsy value: no redirect created
-                },
+                // createRedirects(existingPath) {
+                //     if (!existingPath.endsWith('/')) {
+                //         return `${existingPath}/`;
+                //     }
+                //
+                //     return undefined; // Return a falsy value: no redirect created
+                // },
             },
         ],
         [
@@ -170,6 +172,43 @@ module.exports = {
                 },
             };
         },
+        // skipping svgo for animated crawlee logo
+        async function doNotUseSVGO() {
+            return {
+                name: 'docusaurus-svgo',
+                configureWebpack(config) {
+                    // find the svg rule
+                    const svgRule = config.module.rules.find((r) => typeof r === 'object' && r.test.toString() === '/\\.svg$/i');
+
+                    // find the svgr loader
+                    const svgrLoader = svgRule?.oneOf?.[0];
+
+                    // make copy of svgr loader and disable svgo
+                    const svgrLoaderCopy = JSON.parse(JSON.stringify(svgrLoader));
+
+                    // include only animated logo
+                    svgrLoaderCopy.include = /animated-crawlee-logo/;
+
+                    // turn off svgo
+                    svgrLoaderCopy.use[0].options.svgo = false;
+
+                    // insert the copy after the original svgr loader
+                    svgRule.oneOf.splice(1, 0, svgrLoaderCopy);
+
+                    // exclude animated logo from the first svgr loader (with svgo enabled)
+                    svgrLoader.exclude = /animated-crawlee-logo/;
+
+                    return {
+                        mergeStrategy: {
+                            'module.rules': 'replace',
+                        },
+                        module: {
+                            rules: config.module.rules,
+                        },
+                    };
+                },
+            };
+        },
     ],
     themeConfig:
     /** @type {import('@docusaurus/preset-classic').ThemeConfig} */ ({
@@ -179,10 +218,10 @@ module.exports = {
                 hideable: true,
             },
         },
-        announcementBar: {
-            id: `crawlee-for-python-webinar`,
-            content: `🎉️ <b><a href="https://crawlee.dev/python/">Crawlee for Python is open to early adopters!</a></b> 🥳️`,
-        },
+        // announcementBar: {
+        //     id: `crawlee-for-python-webinar`,
+        //     content: `🎉️ <b><a href="https://crawlee.dev/python/">Crawlee for Python is open to early adopters!</a></b> 🥳️`,
+        // },
         navbar: {
             hideOnScroll: true,
             title: 'Crawlee',
@@ -223,53 +262,6 @@ module.exports = {
                     label: 'Blog',
                     position: 'left',
                 },
-                {
-                    type: 'dropdown',
-                    label: 'Node.js',
-                    position: 'left',
-                    items: [
-                        {
-                            label: 'Node.js',
-                            href: '#',
-                            target: '_self',
-                            rel: 'dofollow',
-                        },
-                        {
-                            label: 'Python',
-                            href: 'https://crawlee.dev/python',
-                            target: '_self',
-                            rel: 'dofollow',
-                        },
-                    ],
-                },
-                {
-                    type: 'docsVersionDropdown',
-                    position: 'left',
-                    dropdownItemsAfter: [
-                        {
-                            href: 'https://sdk.apify.com/docs/guides/getting-started',
-                            label: '2.2',
-                        },
-                        {
-                            href: 'https://sdk.apify.com/docs/1.3.1/guides/getting-started',
-                            label: '1.3',
-                        },
-                    ],
-                },
-                {
-                    href: 'https://github.com/apify/crawlee',
-                    label: 'GitHub',
-                    title: 'View on GitHub',
-                    position: 'right',
-                    className: 'icon',
-                },
-                {
-                    href: 'https://discord.com/invite/jyEM2PRvMU',
-                    label: 'Discord',
-                    title: 'Chat on Discord',
-                    position: 'right',
-                    className: 'icon',
-                },
             ],
         },
         colorMode: {
@@ -297,29 +289,25 @@ module.exports = {
                     items: [
                         {
                             label: 'Guides',
-                            to: 'docs/guides',
+                            to: 'js/docs/guides',
                         },
                         {
                             label: 'Examples',
-                            to: 'docs/examples',
+                            to: 'js/docs/examples',
                         },
                         {
                             label: 'API reference',
-                            to: 'api/core',
+                            to: 'js/api/core',
                         },
                         {
-                            label: 'Upgrading to v3',
-                            to: 'docs/upgrading/upgrading-to-v3',
+                            label: 'Changelog',
+                            to: 'js/api/core/changelog',
                         },
                     ],
                 },
                 {
-                    title: 'Community',
+                    title: 'Product',
                     items: [
-                        {
-                            label: 'Blog',
-                            to: 'blog',
-                        },
                         {
                             label: 'Discord',
                             href: 'https://discord.com/invite/jyEM2PRvMU',
@@ -332,13 +320,18 @@ module.exports = {
                             label: 'Twitter',
                             href: 'https://twitter.com/apify',
                         },
+                        {
+                            label: 'YouTube',
+                            href: 'https://www.youtube.com/apify',
+                        },
+
                     ],
                 },
                 {
                     title: 'More',
                     items: [
                         {
-                            label: 'Apify Platform',
+                            label: 'Apify platform',
                             href: 'https://apify.com',
                         },
                         {
@@ -352,19 +345,19 @@ module.exports = {
                     ],
                 },
             ],
-            logo: {
-                src: 'img/apify_logo.svg',
-                href: '/',
-                width: '60px',
-                height: '60px',
-            },
         },
         algolia: {
             appId: '5JC94MPMLY',
             apiKey: '267679200b833c2ca1255ab276731869', // search only (public) API key
             indexName: 'crawlee',
+            placeholder: 'Search documentation',
             algoliaOptions: {
                 facetFilters: ['version:VERSION'],
+            },
+            translations: {
+                button: {
+                    buttonText: 'Search documentation...',
+                },
             },
         },
     }),