+++ website/versioned_docs/version-3.13/guides/proxy_management_inspection_puppeteer.ts
@@ -1,0 +1,13 @@
+import { PuppeteerCrawler, ProxyConfiguration } from 'crawlee';
+
+const proxyConfiguration = new ProxyConfiguration({
+    /* opts */
+});
+
+const crawler = new PuppeteerCrawler({
+    proxyConfiguration,
+    async requestHandler({ proxyInfo }) {
+        console.log(proxyInfo);
+    },
+    // ...
+});